Query to initialise our recipe database
This SQL file sets up the initial (empty) tables required to store essential recipe information. It includes structures for step-by-step instructions for each recipe and the ingredients associated with each recipe ID. This database will work in conjunction with a fridge contents database (which is yet to be developed) to suggest suitable recipes. While the database is currently empty, it will be populated in a later stage of development.