SQL query for the available ingredients.
The SQL queries that has been added would create a Database with tables which would be populated by the data captured by the SSH camera. The tables include the main table which contains data about the available ingredients (fridge_contents), along with the units table and the ingredients table which are helpful in populating the main table. For testing purposes, we would populate this database with test data.