Update and rename RecipeFilterApp to RecipeFilterApp.java
I have written the function we will use to check the database for recipes that are valid, i.e. recipes that the user is able to make with the contents of their fridge. I have put a temporary SQL command to just select all for now, once Fahad has finished with the actual SQL logic we will use, it can be added in.